Add files via upload
The document consists of detail oriented project in the end user side. The document consists of various questionnaires , facts , personas ,  observation , facts , detail-analysis about it. The project was developed on the testing of the website but on the end -user experience testing.User experience testing is the process of testing different aspects of user experience to determine the best way for a website and its elements to interact with its audience.